refactor: 좋아요 버튼 이벤트 핸들러 분리
